refactor: move initial msfs path setup functions to install path utils
